pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xl"
version = "0.1.0"
description = "XL — Reverse Translation & Folio Structuring (MS Erfurt Phase 1)"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "anthropic>=0.40",
    "openai>=1.40",
    "pillow>=10.0",
    "numpy>=1.24",
    "imagehash>=4.3.0",
]

[project.scripts]
xl = "xl.__main__:main"
scribesim = "scribesim.__main__:main"
weather = "weather.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["xl*", "scribesim*", "weather*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
